+++ griptape/drivers/prompt/google_prompt_driver.py
@@ -1,15 +1,27 @@
 from __future__ import annotations
+
 from collections.abc import Iterator
-from typing import TYPE_CHECKING, Optional, Any
-from attrs import define, field, Factory
-from griptape.utils import PromptStack, import_optional_dependency
+from typing import TYPE_CHECKING, Any, Optional
+
+from attrs import Factory, define, field
+
 from griptape.artifacts import TextArtifact
+from griptape.common import (
+    BaseMessageContent,
+    DeltaMessage,
+    TextDeltaMessageContent,
+    ImageMessageContent,
+    PromptStack,
+    Message,
+    TextMessageContent,
+)
 from griptape.drivers import BasePromptDriver
-from griptape.tokenizers import GoogleTokenizer, BaseTokenizer
+from griptape.tokenizers import BaseTokenizer, GoogleTokenizer
+from griptape.utils import import_optional_dependency
 
 if TYPE_CHECKING:
     from google.generativeai import GenerativeModel
-    from google.generativeai.types import ContentDict
+    from google.generativeai.types import ContentDict, GenerateContentResponse
 
 
 @define
@@ -33,12 +45,12 @@ class GooglePromptDriver(BasePromptDriver):
     top_p: Optional[float] = field(default=None, kw_only=True, metadata={"serializable": True})
     top_k: Optional[int] = field(default=None, kw_only=True, metadata={"serializable": True})
 
-    def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
+    def try_run(self, prompt_stack: PromptStack) -> Message:
         GenerationConfig = import_optional_dependency("google.generativeai.types").GenerationConfig
 
-        inputs = self._prompt_stack_to_model_input(prompt_stack)
-        response = self.model_client.generate_content(
-            inputs,
+        messages = self._prompt_stack_to_messages(prompt_stack)
+        response: GenerateContentResponse = self.model_client.generate_content(
+            messages,
             generation_config=GenerationConfig(
                 stop_sequences=self.tokenizer.stop_sequences,
                 max_output_tokens=self.max_tokens,
@@ -48,14 +60,22 @@ class GooglePromptDriver(BasePromptDriver):
             ),
         )
 
-        return TextArtifact(value=response.text)
+        usage_metadata = response.usage_metadata
 
-    def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]:
+        return Message(
+            content=[TextMessageContent(TextArtifact(response.text))],
+            role=Message.ASSISTANT_ROLE,
+            usage=Message.Usage(
+                input_tokens=usage_metadata.prompt_token_count, output_tokens=usage_metadata.candidates_token_count
+            ),
+        )
+
+    def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
         GenerationConfig = import_optional_dependency("google.generativeai.types").GenerationConfig
 
-        inputs = self._prompt_stack_to_model_input(prompt_stack)
-        response = self.model_client.generate_content(
-            inputs,
+        messages = self._prompt_stack_to_messages(prompt_stack)
+        response: Iterator[GenerateContentResponse] = self.model_client.generate_content(
+            messages,
             stream=True,
             generation_config=GenerationConfig(
                 stop_sequences=self.tokenizer.stop_sequences,
@@ -66,16 +86,27 @@ class GooglePromptDriver(BasePromptDriver):
             ),
         )
 
+        prompt_token_count = None
         for chunk in response:
-            yield TextArtifact(value=chunk.text)
+            usage_metadata = chunk.usage_metadata
 
-    def _prompt_stack_input_to_message(self, prompt_input: PromptStack.Input) -> dict:
-        parts = [prompt_input.content]
-
-        if prompt_input.is_assistant():
-            return {"role": "model", "parts": parts}
-        else:
-            return {"role": "user", "parts": parts}
+            # Only want to output the prompt token count once since it is static each chunk
+            if prompt_token_count is None:
+                prompt_token_count = usage_metadata.prompt_token_count
+                yield DeltaMessage(
+                    content=TextDeltaMessageContent(chunk.text),
+                    role=Message.ASSISTANT_ROLE,
+                    usage=DeltaMessage.Usage(
+                        input_tokens=usage_metadata.prompt_token_count,
+                        output_tokens=usage_metadata.candidates_token_count,
+                    ),
+                )
+            else:
+                yield DeltaMessage(
+                    content=TextDeltaMessageContent(chunk.text),
+                    role=Message.ASSISTANT_ROLE,
+                    usage=DeltaMessage.Usage(output_tokens=usage_metadata.candidates_token_count),
+                )
 
     def _default_model_client(self) -> GenerativeModel:
         genai = import_optional_dependency("google.generativeai")
@@ -83,20 +114,35 @@ class GooglePromptDriver(BasePromptDriver):
 
         return genai.GenerativeModel(self.model)
 
-    def _prompt_stack_to_model_input(self, prompt_stack: PromptStack) -> list[ContentDict]:
+    def _prompt_stack_to_messages(self, prompt_stack: PromptStack) -> list[dict]:
         inputs = [
-            self.__to_content_dict(prompt_input) for prompt_input in prompt_stack.inputs if not prompt_input.is_system()
+            {"role": self.__to_role(message), "parts": self.__to_content(message)}
+            for message in prompt_stack.messages
+            if not message.is_system()
         ]
 
         # Gemini does not have the notion of a system message, so we insert it as part of the first message in the history.
-        system = next((i for i in prompt_stack.inputs if i.is_system()), None)
-        if system is not None:
-            inputs[0]["parts"].insert(0, system.content)
+        system_messages = prompt_stack.system_messages
+        if system_messages:
+            inputs[0]["parts"].insert(0, system_messages[0].to_text())
 
         return inputs
 
-    def __to_content_dict(self, prompt_input: PromptStack.Input) -> ContentDict:
+    def __prompt_stack_content_message_content(self, content: BaseMessageContent) -> ContentDict | str:
         ContentDict = import_optional_dependency("google.generativeai.types").ContentDict
-        message = self._prompt_stack_input_to_message(prompt_input)
 
-        return ContentDict(message)
+        if isinstance(content, TextMessageContent):
+            return content.artifact.to_text()
+        elif isinstance(content, ImageMessageContent):
+            return ContentDict(mime_type=content.artifact.mime_type, data=content.artifact.value)
+        else:
+            raise ValueError(f"Unsupported content type: {type(content)}")
+
+    def __to_role(self, message: Message) -> str:
+        if message.is_assistant():
+            return "model"
+        else:
+            return "user"
+
+    def __to_content(self, message: Message) -> list[ContentDict | str]:
+        return [self.__prompt_stack_content_message_content(content) for content in message.content]