+++ griptape/engines/extraction/json_extraction_engine.py
@@ -3,9 +3,10 @@ from typing import Optional, cast
 import json
 from attrs import field, Factory, define
 from griptape.artifacts import TextArtifact, ListArtifact, ErrorArtifact
+from griptape.common.prompt_stack.messages.message import Message
 from griptape.engines import BaseExtractionEngine
 from griptape.utils import J2
-from griptape.utils import PromptStack
+from griptape.common import PromptStack
 from griptape.rules import Ruleset
 
 
@@ -58,9 +59,7 @@ class JsonExtractionEngine(BaseExtractionEngine):
         if self.prompt_driver.tokenizer.count_input_tokens_left(full_text) >= self.min_response_tokens:
             extractions.extend(
                 self.json_to_text_artifacts(
-                    self.prompt_driver.run(
-                        PromptStack(inputs=[PromptStack.Input(full_text, role=PromptStack.USER_ROLE)])
-                    ).value
+                    self.prompt_driver.run(PromptStack(messages=[Message(full_text, role=Message.USER_ROLE)])).value
                 )
             )
 
@@ -75,9 +74,7 @@ class JsonExtractionEngine(BaseExtractionEngine):
 
             extractions.extend(
                 self.json_to_text_artifacts(
-                    self.prompt_driver.run(
-                        PromptStack(inputs=[PromptStack.Input(partial_text, role=PromptStack.USER_ROLE)])
-                    ).value
+                    self.prompt_driver.run(PromptStack(messages=[Message(partial_text, role=Message.USER_ROLE)])).value
                 )
             )