+++ griptape/engines/rag/rag_engine.py
@@ -1,17 +1,34 @@
 from typing import Optional
 from attrs import define, field
 from griptape.engines.rag import RagContext
-from griptape.engines.rag.stages import QueryRagStage, GenerationRagStage, RetrievalRagStage
+from griptape.engines.rag.stages import QueryRagStage, ResponseRagStage, RetrievalRagStage
 
 
 @define(kw_only=True)
 class RagEngine:
     query_stage: Optional[QueryRagStage] = field(default=None)
     retrieval_stage: Optional[RetrievalRagStage] = field(default=None)
-    generation_stage: Optional[GenerationRagStage] = field(default=None)
+    response_stage: Optional[ResponseRagStage] = field(default=None)
+
+    def __attrs_post_init__(self) -> None:
+        modules = []
+
+        if self.query_stage is not None:
+            modules.extend(self.query_stage.modules)
+
+        if self.retrieval_stage is not None:
+            modules.extend(self.retrieval_stage.modules)
+
+        if self.response_stage is not None:
+            modules.extend(self.response_stage.modules)
+
+        module_names = [m.name for m in modules]
+
+        if len(module_names) > len(set(module_names)):
+            raise ValueError("module names have to be unique")
 
     def process_query(self, query: str) -> RagContext:
-        return self.process(RagContext(initial_query=query))
+        return self.process(RagContext(query=query))
 
     def process(self, context: RagContext) -> RagContext:
         if self.query_stage:
@@ -20,7 +37,7 @@ class RagEngine:
         if self.retrieval_stage:
             context = self.retrieval_stage.run(context)
 
-        if self.generation_stage:
-            context = self.generation_stage.run(context)
+        if self.response_stage:
+            context = self.response_stage.run(context)
 
         return context