+++ griptape/tasks/prompt_task.py
@@ -1,10 +1,15 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, Optional, Callable
-from attrs import define, field, Factory
-from griptape.utils import PromptStack
-from griptape.utils import J2
-from griptape.tasks import BaseTextInputTask
+
+from typing import TYPE_CHECKING, Callable, Optional
+
+from attrs import Factory, define, field
+
 from griptape.artifacts import BaseArtifact
+from griptape.common import PromptStack
+from griptape.tasks import BaseTask
+from griptape.utils import J2
+from griptape.artifacts import TextArtifact, ListArtifact
+from griptape.mixins import RuleMixin
 
 if TYPE_CHECKING:
     from griptape.drivers import BasePromptDriver
@@ -12,11 +17,23 @@ if TYPE_CHECKING:
 
 
 @define
-class PromptTask(BaseTextInputTask):
+class PromptTask(RuleMixin, BaseTask):
     _prompt_driver: Optional[BasePromptDriver] = field(default=None, kw_only=True, alias="prompt_driver")
     generate_system_template: Callable[[PromptTask], str] = field(
         default=Factory(lambda self: self.default_system_template_generator, takes_self=True), kw_only=True
     )
+    _input: str | list | tuple | BaseArtifact | Callable[[BaseTask], BaseArtifact] = field(
+        default=lambda task: task.full_context["args"][0] if task.full_context["args"] else TextArtifact(value=""),
+        alias="input",
+    )
+
+    @property
+    def input(self) -> BaseArtifact:
+        return self._process_task_input(self._input)
+
+    @input.setter
+    def input(self, value: str | list | tuple | BaseArtifact | Callable[[BaseTask], BaseArtifact]) -> None:
+        self._input = value
 
     output: Optional[BaseArtifact] = field(default=None, init=False)
 
@@ -25,12 +42,12 @@ class PromptTask(BaseTextInputTask):
         stack = PromptStack()
         memory = self.structure.conversation_memory
 
-        stack.add_system_input(self.generate_system_template(self))
+        stack.add_system_message(self.generate_system_template(self))
 
-        stack.add_user_input(self.input.to_text())
+        stack.add_user_message(self.input)
 
         if self.output:
-            stack.add_assistant_input(self.output.to_text())
+            stack.add_assistant_message(self.output)
 
         if memory:
             # inserting at index 1 to place memory right after system prompt
@@ -59,7 +76,33 @@ class PromptTask(BaseTextInputTask):
             rulesets=J2("rulesets/rulesets.j2").render(rulesets=self.all_rulesets)
         )
 
-    def run(self) -> BaseArtifact:
-        self.output = self.prompt_driver.run(self.prompt_stack)
+    def before_run(self) -> None:
+        super().before_run()
 
-        return self.output
+        self.structure.logger.info(f"{self.__class__.__name__} {self.id}\nInput: {self.input.to_text()}")
+
+    def after_run(self) -> None:
+        super().after_run()
+
+        self.structure.logger.info(f"{self.__class__.__name__} {self.id}\nOutput: {self.output.to_text()}")
+
+    def run(self) -> BaseArtifact:
+        message = self.prompt_driver.run(self.prompt_stack)
+
+        return message.to_artifact()
+
+    def _process_task_input(
+        self, task_input: str | tuple | list | BaseArtifact | Callable[[BaseTask], BaseArtifact]
+    ) -> BaseArtifact:
+        if isinstance(task_input, TextArtifact):
+            task_input.value = J2().render_from_string(task_input.value, **self.full_context)
+
+            return task_input
+        elif isinstance(task_input, Callable):
+            return self._process_task_input(task_input(self))
+        elif isinstance(task_input, BaseArtifact):
+            return task_input
+        elif isinstance(task_input, list) or isinstance(task_input, tuple):
+            return ListArtifact([self._process_task_input(elem) for elem in task_input])
+        else:
+            return self._process_task_input(TextArtifact(task_input))