+++ tests/unit/drivers/prompt/test_hugging_face_pipeline_prompt_driver.py
@@ -1,5 +1,5 @@
 from griptape.drivers import HuggingFacePipelinePromptDriver
-from griptape.utils import PromptStack
+from griptape.common import PromptStack
 import pytest
 
 
@@ -22,15 +22,15 @@ class TestHuggingFacePipelinePromptDriver:
         mock_autotokenizer.model_max_length = 42
         mock_autotokenizer.apply_chat_template.return_value = [1, 2, 3]
         mock_autotokenizer.decode.return_value = "model-output"
+        mock_autotokenizer.encode.return_value = [1, 2, 3]
         return mock_autotokenizer
 
     @pytest.fixture
     def prompt_stack(self):
         prompt_stack = PromptStack()
-        prompt_stack.add_generic_input("generic-input")
-        prompt_stack.add_system_input("system-input")
-        prompt_stack.add_user_input("user-input")
-        prompt_stack.add_assistant_input("assistant-input")
+        prompt_stack.add_system_message("system-input")
+        prompt_stack.add_user_message("user-input")
+        prompt_stack.add_assistant_message("assistant-input")
         return prompt_stack
 
     def test_init(self):
@@ -41,10 +41,12 @@ class TestHuggingFacePipelinePromptDriver:
         driver = HuggingFacePipelinePromptDriver(model="foo", max_tokens=42)
 
         # When
-        text_artifact = driver.try_run(prompt_stack)
+        message = driver.try_run(prompt_stack)
 
         # Then
-        assert text_artifact.value == "model-output"
+        assert message.value == "model-output"
+        assert message.usage.input_tokens == 3
+        assert message.usage.output_tokens == 3
 
     def test_try_stream(self, prompt_stack):
         # Given