+++ tests/unit/engines/rag/modules/test_base_rag_nodule.py
@@ -1,0 +1,27 @@
+from griptape.engines.rag import RagContext
+from tests.mocks.mock_rag_module import MockRagModule
+
+
+class TestBaseRagModule:
+    def test_generate_query_prompt_stack(self):
+        prompt_stack = MockRagModule().generate_query_prompt_stack("test system", "test query")
+
+        assert len(prompt_stack.messages) == 2
+        assert prompt_stack.messages[0].is_system()
+        assert prompt_stack.messages[1].is_user()
+
+    def test_get_context_param(self):
+        module = MockRagModule(name="boo")
+        context = RagContext(query="test")
+
+        context.module_params["boo"] = {"foo": "bar"}
+
+        assert module.get_context_param(context, "foo") == "bar"
+
+    def test_set_context_param(self):
+        module = MockRagModule(name="boo")
+        context = RagContext(query="test")
+
+        module.set_context_param(context, "foo", "bar")
+
+        assert context.module_params["boo"]["foo"] == "bar"