+++ tests/unit/structures/test_workflow.py
@@ -351,6 +351,44 @@ class TestWorkflow:
 
         self._validate_topology_1(workflow)
 
+    def test_run_topology_1_missing_parent(self):
+        task1 = PromptTask("test1", id="task1")
+        task2 = PromptTask("test2", id="task2")
+        task3 = PromptTask("test3", id="task3")
+        task4 = PromptTask("test4", id="task4")
+        workflow = Workflow(prompt_driver=MockPromptDriver())
+
+        # task1 never added to workflow
+        workflow + task4
+        with pytest.raises(ValueError):
+            workflow.insert_tasks(task1, [task2, task3], task4)
+
+    def test_run_topology_1_id_equality(self):
+        task1 = PromptTask("test1", id="task1")
+        task2 = PromptTask("test2", id="task2")
+        task3 = PromptTask("test3", id="task3")
+        task4 = PromptTask("test4", id="task4")
+        workflow = Workflow(prompt_driver=MockPromptDriver())
+
+        # task4 never added to workflow
+        workflow + task1
+        workflow.insert_tasks(task1, [task2, task3], task4)
+
+        with pytest.raises(ValueError):
+            workflow.run()
+
+    def test_run_topology_1_object_equality(self):
+        task1 = PromptTask("test1", id="task1")
+        task2 = PromptTask("test2", id="task2")
+        task3 = PromptTask("test3", id="task3")
+        task4 = PromptTask("test4", id="task4")
+        workflow = Workflow(prompt_driver=MockPromptDriver())
+
+        workflow + task1
+        workflow + task4
+        with pytest.raises(ValueError):
+            workflow.insert_tasks(PromptTask("test1", id="task1"), [task2, task3], task4)
+
     def test_run_topology_2_declarative_parents(self):
         workflow = Workflow(
             prompt_driver=MockPromptDriver(),
@@ -645,7 +683,7 @@ class TestWorkflow:
         # task4 is the final task, but its defined at index 0
         workflow = Workflow(prompt_driver=MockPromptDriver(), tasks=[task4, task1, task2, task3])
 
-        # ouput_task topologically should be task4
+        # output_task topologically should be task4
         assert task4 == workflow.output_task
 
     def test_to_graph(self):