+++ tests/unit/tasks/test_toolkit_task.py
@@ -2,7 +2,7 @@ from griptape.artifacts import ErrorArtifact, TextArtifact
 from griptape.structures import Agent
 from griptape.tasks import ToolkitTask, ActionsSubtask, PromptTask
 from tests.mocks.mock_tool.tool import MockTool
-from tests.mocks.mock_value_prompt_driver import MockValuePromptDriver
+from tests.mocks.mock_prompt_driver import MockPromptDriver
 from tests.utils import defaults
 
 
@@ -149,7 +149,7 @@ class TestToolkitSubtask:
         output = """Answer: done"""
 
         task = ToolkitTask("test", tools=[MockTool(name="Tool1"), MockTool(name="Tool2")])
-        agent = Agent(prompt_driver=MockValuePromptDriver(value=output))
+        agent = Agent(prompt_driver=MockPromptDriver(mock_output=output))
 
         agent.add_task(task)
 
@@ -163,7 +163,7 @@ class TestToolkitSubtask:
         output = """Actions: [{"name": "blah"}]"""
 
         task = ToolkitTask("test", tools=[MockTool(name="Tool1")], max_subtasks=3)
-        agent = Agent(prompt_driver=MockValuePromptDriver(value=output))
+        agent = Agent(prompt_driver=MockPromptDriver(mock_output=output))
 
         agent.add_task(task)
 
@@ -176,7 +176,7 @@ class TestToolkitSubtask:
         output = """foo bar"""
 
         task = ToolkitTask("test", tools=[MockTool(name="Tool1")], max_subtasks=3)
-        agent = Agent(prompt_driver=MockValuePromptDriver(value=output))
+        agent = Agent(prompt_driver=MockPromptDriver(mock_output=output))
 
         agent.add_task(task)